aoe damage
does damage to player if player collider collides with whatever collider the EnemyAoE script is attached to. it will also decrease the player's health (and the player health script should be attached to the player collider but that might be the wrong place to put it, sorry idk unity) by 1 every 0.5 seconds until the player leaves the collider. the DamageOverTime coroutine will stop when the player leaves the collider and if it was ever running to begin with.