health bar works
To do:
- fix health bar flipping bug. cannot just health bar above player because it then starts removing the left side of the bar
- add a sprite to fill the health bar and also a border

Health bar is a red rectangle that follows player, so when player flips to turn the other way, the health bar also flips sides (to be fixed).

Health bar is a slider object, and its value is determined by its health. The amount of health in it is the same as the player's health. In the player collider, the PlayerHealth script takes in a HealthBar and transfers the PlayerHealth to the HealthBar. The HealthBar script is attached to the once empty object with a slider and fill attached to it. This fill has to be replaced with an actual image and optionally add a border later